replace preloader and prettify adapter
